perf: statically generate event detail page with 60s ISR
Server-render /events/[eventId] with generateStaticParams for the top
20 events by attendance; move interactivity (hero actions, tabs,
ticket purchase) into client components.
