+++ src/mocks/events.ts
@@ -12,11 +12,14 @@ export const mockEvents: Event[] = [
     date: "June 15-17, 2024",
     dateEnd: "June 17, 2024",
     time: "6:00 PM - 2:00 AM",
+    eventDate: "2024-06-15",
     location: "Central Park, New York",
     venue: "Central Park",
     price: "0.08 ETH",
     priceInEth: 0.08,
     category: "festival",
+    status: "PUBLISHED",
+    organizerId: "rhythm-nation-collective",
     featured: true,
     attendees: 1200,
     description:
@@ -65,11 +68,14 @@ export const mockEvents: Event[] = [
     image: "/images/events/detail2.png",
     date: "July 8, 2024",
     time: "9:00 PM - 4:00 AM",
+    eventDate: "2024-07-08",
     location: "Central Park, New York",
     venue: "Central Park",
     price: "0.05 ETH",
     priceInEth: 0.05,
     category: "music",
+    status: "PUBLISHED",
+    organizerId: "beat-collective",
     featured: false,
     attendees: 800,
     description:
@@ -86,11 +92,14 @@ export const mockEvents: Event[] = [
     image: "/images/events/detail.png",
     date: "June 15-17, 2024",
     time: "6:00 PM - 2:00 AM",
+    eventDate: "2024-06-15",
     location: "Central Park, New York",
     venue: "Central Park",
     price: "0.08 ETH",
     priceInEth: 0.08,
     category: "festival",
+    status: "PUBLISHED",
+    organizerId: "rhythm-nation-collective",
     featured: false,
     attendees: 1200,
   },
@@ -100,11 +109,14 @@ export const mockEvents: Event[] = [
     image: "/images/events/detail2.png",
     date: "July 8, 2024",
     time: "9:00 PM - 4:00 AM",
+    eventDate: "2024-07-08",
     location: "Central Park, New York",
     venue: "Central Park",
     price: "0.05 ETH",
     priceInEth: 0.05,
     category: "music",
+    status: "PUBLISHED",
+    organizerId: "beat-collective",
     featured: false,
     attendees: 800,
   },
